陽性者の属性: Rename components/cards/ConfirmedCasesAttributesCard.vue to components/index/CardsReference/ConfirmedCasesAttributes/Card.vue
