+++ components/index/CardsReference/ConfirmedCasesAttributes/Card.vue
